Write first passing test
- Mock canvas
- Remove App file and mount/unmount ReactDOM on import
