+++ src/__tests__/mouseDown.test.tsx
@@ -1,7 +1,15 @@
 import React from "react";
-import { App } from "../App";
+import ReactDOM from "react-dom";
+import { render, fireEvent } from "@testing-library/react";
+import { App } from "../index";
+
+// Unmount ReactDOM from root
+ReactDOM.unmountComponentAtNode(document.getElementById("root")!);
 
 it("just works", () => {
-  // eslint-disable-next-line
-  const app = <App />;
+  const { getByTitle } = render(<App />);
+  const rectangleTool = getByTitle("Rectangle — R, 2");
+  const radio = rectangleTool.querySelector("input") as HTMLInputElement;
+  fireEvent.click(rectangleTool);
+  expect(radio.checked).toBeTruthy();
 });
