+++ src/setupTests.ts
@@ -1,1 +1,8 @@
 import "@testing-library/jest-dom";
+import "jest-canvas-mock";
+
+// ReactDOM is located inside index.tsx file
+// as a result, we need a place for it to render into
+const element = document.createElement("div");
+element.id = "root";
+document.body.appendChild(element);
